functions/auth_callback/pyproject.toml
[project]
name = "auth-callback"
version = "0.1.0"
description = "GitHub App OAuth callback Cloud Function"
requires-python = ">=3.12"
dependencies = [
    "functions-framework>=3.0,<5",
    "functions-shared",
    "google-cloud-firestore>=2.0,<4",
]

[tool.uv.sources]
functions-shared = { workspace = true }

[tool.setuptools]
py-modules = ["main"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
